requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.10.0
newspaper3k>=0.2.8
lxml[html_clean]
tqdm